+++ jannovar-cli/src/main/java/de/charite/compbio/jannovar/cmd/db_list/DatabaseListCommand.java
@@ -1,19 +1,20 @@
 package de.charite.compbio.jannovar.cmd.db_list;
 
-import org.apache.commons.cli.ParseException;
-
 import de.charite.compbio.jannovar.JannovarException;
-import de.charite.compbio.jannovar.JannovarOptions;
 import de.charite.compbio.jannovar.cmd.CommandLineParsingException;
-import de.charite.compbio.jannovar.cmd.HelpRequestedException;
 import de.charite.compbio.jannovar.cmd.JannovarCommand;
 import de.charite.compbio.jannovar.datasource.DataSourceFactory;
 import de.charite.compbio.jannovar.datasource.DatasourceOptions;
+import net.sourceforge.argparse4j.inf.Namespace;
 
 public class DatabaseListCommand extends JannovarCommand {
 
-	public DatabaseListCommand(String[] argv) throws CommandLineParsingException, HelpRequestedException {
-		super(argv);
+	/** Configuration */
+	private JannovarDBListOptions options;
+
+	public DatabaseListCommand(String argv[], Namespace args) throws CommandLineParsingException {
+		this.options = new JannovarDBListOptions();
+		this.options.setFromArgs(args);
 	}
 
 	/**
@@ -22,25 +23,15 @@ public class DatabaseListCommand extends JannovarCommand {
 	@Override
 	public void run() throws JannovarException {
 		System.err.println("Options");
-		options.print(System.err);
+		System.err.println(options.toString());
 
-		DatasourceOptions dsOptions = new DatasourceOptions(options.httpProxy, options.httpsProxy, options.ftpProxy,
-				options.printProgressBars);
+		DatasourceOptions dsOptions = new DatasourceOptions(options.getHttpProxy(), options.getHttpsProxy(),
+				options.getFtpProxy(), options.isReportProgress());
 
-		DataSourceFactory factory = new DataSourceFactory(dsOptions, options.dataSourceFiles);
+		DataSourceFactory factory = new DataSourceFactory(dsOptions, options.getDataSourceFiles());
 		System.err.println("Available data sources:\n");
 		for (String name : factory.getNames())
 			System.err.println(String.format("    %s", name));
-	}
-
-	@Override
-	protected JannovarOptions parseCommandLine(String[] argv) throws CommandLineParsingException,
-			HelpRequestedException {
-		try {
-			return new DatabaseListCommandLineParser().parse(argv);
-		} catch (ParseException e) {
-			throw new CommandLineParsingException("Could not parse command line", e);
-		}
 	}
 
 }