+++ jannovar-cli/src/test/java/de/charite/compbio/jannovar/cmd/annotate_vcf/JannovarAnnotateVCFTest.java
@@ -14,6 +14,7 @@ import com.google.common.base.Charsets;
 import com.google.common.base.Joiner;
 import com.google.common.io.Files;
 
+import de.charite.compbio.jannovar.Jannovar;
 import de.charite.compbio.jannovar.JannovarException;
 
 /**
@@ -37,10 +38,12 @@ public class JannovarAnnotateVCFTest {
 	public void testOnSmallExample() throws JannovarException, URISyntaxException, IOException {
 		final File outFolder = tmpFolder.newFolder();
 		final String inputFilePath = this.getClass().getResource("/small.vcf").toURI().getPath();
-		String[] argv = new String[] { "annotate", "-o", outFolder.toString(), "-d", pathToSmallSer, "-i",
-				inputFilePath };
+		String[] argv = new String[] { "annotate-vcf", "-o", outFolder.toString() + "/small.jv.vcf", "-d",
+				pathToSmallSer, "-i", inputFilePath };
 		System.err.println(Joiner.on(" ").join(argv));
-		new AnnotateVCFCommand(argv).run();
+
+		Jannovar.main(argv);
+
 		File f = new File(outFolder.getAbsolutePath() + File.separator + "small.jv.vcf");
 		Assert.assertTrue(f.exists());
 
@@ -57,10 +60,12 @@ public class JannovarAnnotateVCFTest {
 	public void testOnTrailingSemicolons() throws JannovarException, URISyntaxException, IOException {
 		final File outFolder = tmpFolder.newFolder();
 		final String inputFilePath = this.getClass().getResource("/semicolons.vcf").toURI().getPath();
-		String[] argv = new String[] { "annotate", "-o", outFolder.toString(), "-d", pathToSmallSer, "-i",
-				inputFilePath };
+		String[] argv = new String[] { "annotate-vcf", "-o", outFolder.toString() + "/semicolons.jv.vcf", "-d",
+				pathToSmallSer, "-i", inputFilePath };
 		System.err.println(Joiner.on(" ").join(argv));
-		new AnnotateVCFCommand(argv).run();
+
+		Jannovar.main(argv);
+
 		File f = new File(outFolder.getAbsolutePath() + File.separator + "semicolons.jv.vcf");
 		Assert.assertTrue(f.exists());